pyproject.toml
[project]
name = "deepseek-qwen-distill"
version = "0.0.1"
description = "DeepSeek-R1 Distillation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "markupsafe==2.1.3",
    "torch==2.4.1",
    "transformers>=4.48.3",
    "trl>=0.10.1",
    "peft>=0.14.0",
    "datasets>=2.14.4",
    "python-dotenv>=1.0.1",
]


[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu121"
